Initial set of changes for multiple robots in director
applogic.py
-----------

Modify removeToolBarMacro to use .get function of dict instead of trying to access the name directly. This means that if you try to remove a toolbar twice it will just carry on instead of crashing. Maybe this was the intended functionality? In which case the function should return something to give information if a toolbar was already deleted.

cameraview.py
-------------

Modify to take robotname as a parameter to the init function which allows it to initialise without crashing when directorconfig contains more than one robot.

ImageManager and CameraView also take into account the possibility of multiple robots on initialisation when getting the cameraMode and monoCameras keys. The current implementation means that only the values from the first robot in the dict are taken into account.

footstepsdriver.py
------------------

Change the globals for footmeshes etc to live inside the driver class. When a footstepsdriver class is initialised it is added to a global dict which can be accessed with getfootstepsdriver(robotname) which should facilitate the use of existing drivers if needed. Some of the global functions changed to make use of this already.

Change how directorconfig is used for smooth operation with multiple robots.

Changed some of the static methods which used the globals to now be non-static. They can still be used in the same way by calling the getfootstepsdriver global function.

jointcontrol.py
---------------

Makes use of the new getrobotstate function to get the state for a specific robot rather than calling its functions directly.

perception.py
-------------

Minor changes to be able to read parameters from director configs which specify multiple robots.

Specify a prefix for the sensors directory which it creates to separate the sensors of multiple robots (only a partial implementation)

robotstate.py
-------------

Convert the whole thing to a class, make it work with multiple robots, and remove some functions which used lcm things

robotsystem.py
--------------

Changes for various parts of the initialisation to pass the robot name into things which needed it, mostly to grab initialisation parameters from the director configuration file.

Pass the robotname into the factory create function so that it is accessible to the various init functions, and also as one of the fields in the robotSystem object returned

roboturdf.py
------------

Changes to allow reading correct parts of the configuration with multiple robots. Passes the robot name to the jointcontroller initialisation

segmentationroutines.py
-----------------------

Instead of having only a single heightprovider and viewprovider, the segmentationcontext can now have multiple view providers, which allows the use of multiple robots. Not sure if this is actually the intended usage but it makes the system start which is good enough for now.

Modified the initialisation of the context as well, so that the init functions call initcontext which is the thing that actually does the installglobalinstance call.

A change to use getfootstepsdriver with a robot name for the robotmodelgroundheightprovider

startup.py
----------

Small change to how the directorconfig is used, then put everything which is used to initialise a robot system into a loop.