Bad fix of calls to segmentation functions with no reference to specific view provider. Fixes tests
Instead of erroring, if there is only one key in the dict, functions return the value associated with that key. This helps with testing where there is only one key, but in the case of multiple robots the names of the providers should be passed to the function. If there is more than one key the function will error.